pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "endo-io"
version = "0.1.0"
description = "Prototype histogram classifier for endoscope inside vs outside frames"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
endo-io = "endo_io.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
